feat: Add pip-tool management of requirements
In most openedx projects we use a pattern of having requirements.in
files that generate the final requirements.txt using pip-tools. This is
a good pattern and we should continue to use that here.

A Makefile that allows for the installation of requirements.txt and a
list of requirements can be found in requirements.in.