feat: Switch edx-theme for rtd-theme
Given that the edx theme is not actually the right theme for us, we are
not going to use it. Since the default is a bit sparse, I have elected to
use the Read the docs theme as it supports the addition of a logo and
customization of colors.

In addition, the requirements have been updated to reflect the removal
of the edx-theme and the addition of the rtd-theme.