feat: Add first-steps content and style (#5)
* feat: Add first-steps content and style

A module with the first steps has been added via first-steps.rst. This
is references in index.rst. To add better style, the edx-theme has been
added in conf.py.

* feat: Add link to Jira/Confluence

Adding a link so that new users will know hwere to go to register a new
account.

* feat: Add links where appropriate

Links to discuss.openedx.org and github settings have been added.

* feat: Switch edx-theme for rtd-theme

Given that the edx theme is not actually the right theme for us, we are
not going to use it. Since the default is a bit sparse, I have elected to
use the Read the docs theme as it supports the addition of a logo and
customization of colors.

In addition, the requirements have been updated to reflect the removal
of the edx-theme and the addition of the rtd-theme.

* fix: Add SSO to Jira/Confluence

* fix: Replace Authorization with Authentication